+++ BioBloomCategorizer/ResultsManager.h
@@ -23,13 +23,13 @@ public:
 	explicit ResultsManager(const vector<string> &hashSigsRef,
 			const unordered_map<string, shared_ptr<MultiFilter> > &filtersRef,
 			const unordered_map<string, vector<shared_ptr<BloomFilterInfo> > > &infoFilesRef,
-			float scoreThreshold);
+			double scoreThreshold);
 
 	const string updateSummaryData(size_t seqLen,
-			unordered_map<string, float> &hits);
+			unordered_map<string, double> &hits);
 	const string updateSummaryData(size_t seqLen1, size_t seqLen2,
-			unordered_map<string, float> &hits1,
-			unordered_map<string, float> &hits2);
+			unordered_map<string, double> &hits1,
+			unordered_map<string, double> &hits2);
 
 	const string getResultsSummary(size_t readCount) const;
 	virtual ~ResultsManager();
@@ -38,7 +38,7 @@ private:
 	const unordered_map<string, vector<shared_ptr<BloomFilterInfo> > > infoFiles;
 	const unordered_map<string, shared_ptr<MultiFilter> > filters;
 	const vector<string> hashSigs;
-	const float scoreThreshold;
+	const double scoreThreshold;
 
 	unordered_map<string, size_t> aboveThreshold;
 	unordered_map<string, size_t> belowThreshold;
